Adds shared context "JSON conversion"
