+++ spec/fortnox/api/mappers/base/to_json_spec.rb
@@ -1,63 +1,9 @@
 require 'spec_helper'
-require 'fortnox/api'
-require 'fortnox/api/mappers'
 require 'fortnox/api/mappers/base/to_json'
+require 'fortnox/api/mappers/contexts/json_conversion'
 
 describe Fortnox::API::Mapper::ToJSON do
-  before(:all) do
-    module Test
-      class CategoryMapper < Fortnox::API::Mapper::Base
-      end
-      class ProductDesignerMapper < Fortnox::API::Mapper::Base
-      end
-      class ProductMapper < Fortnox::API::Mapper::Base
-      end
-    end
-
-    Fortnox::API::Registry.register( :category, Test::CategoryMapper )
-    Fortnox::API::Registry.register( :categories, Test::CategoryMapper )
-    Fortnox::API::Registry.register( :designer, Test::ProductDesignerMapper )
-    Fortnox::API::Registry.register( :product, Test::ProductMapper )
-  end
-
-  before do
-    module Test
-      class CategoryMapper < Fortnox::API::Mapper::Base
-        KEY_MAP = { id: 'ID' }.freeze
-      end
-
-      class ProductDesignerMapper < Fortnox::API::Mapper::Base
-        KEY_MAP = { id: 'ID' }.freeze
-      end
-
-      class ProductMapper < Fortnox::API::Mapper::Base
-        KEY_MAP = {
-          vat: 'VAT',
-          url: '@url' # TODO: How to handle url attribute?
-        }.freeze
-        JSON_ENTITY_WRAPPER = 'Product'.freeze
-        JSON_COLLECTION_WRAPPER = 'Products'.freeze
-      end
-
-      class Category < Dry::Struct
-        attribute :name, 'strict.string'
-        attribute :id, 'strict.string'
-      end
-
-      class ProductDesigner < Dry::Struct
-        attribute :name, 'strict.string'
-        attribute :id, 'strict.string'
-      end
-
-      class Product < Dry::Struct
-        attribute :url, 'strict.string'
-        attribute :name, 'strict.string'
-        attribute :vat, 'strict.float'
-        attribute :categories, Dry::Types['coercible.array'].member( Test::Category )
-        attribute :designer, Test::ProductDesigner
-      end
-    end
-  end
+  include_context 'JSON conversion'
 
   let( :mapper ){ Test::ProductMapper.new }
